socialin/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.0.10
django-cors-headers==3.14.0
django-environ==0.10.0
django-multiselectfield==0.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.3
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
